requirements.txt
crewai
litellm
python-dotenv
fastapi
uvicorn[standard]
python-multipart
aiofiles
openai
librosa
numpy
